Options for e-mailing errors and an e-mail class.
Set whether to e-mail errors or not by setting
"EMailErrors" to 1 or 0 in "[ErrorHandling]"-section in the
application's configuration file. The default value is false.
Set what e-mail to send errors from by setting "EMailFrom" in
"[ErrorHandling]"-section in the application's configuration file. The
default value is "". Set what e-mail to send errors to by setting
"EMailTo" in "[ErrorHandling]"-section in the application's
configuration file. The
default value is "". Added a class for sending e-mails.